pyproject.toml
[project]
name = "hotelflowbackend"
version = "0.1.0"
description = ""
authors = [
    {name = "defuzia",email = "zhyvhyn.roman@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.13,<4"
dependencies = [
    "django (>=5.2.1,<6.0.0)",
    "djangorestframework (>=3.16.0,<4.0.0)",
    "django-environ (>=0.12.0,<0.13.0)",
    "django-cors-headers (>=4.7.0,<5.0.0)",
    "phonenumbers (>=9.0.5,<10.0.0)",
    "pillow (>=11.2.1,<12.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "qrcode (>=8.2,<9.0)",
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
